+++ packages/popover/src/index.ts
@@ -1,7 +1,8 @@
 import Popover from './Popover';
 import Align from './Align';
 import Justify from './Justify';
+import Justification from './Justification';
 export { PopoverProps } from './Popover';
 
 export default Popover;
-export { Align, Justify };
+export { Align, Justify, Justification };
